+++ built_value/lib/built_value.dart
@@ -4,8 +4,6 @@
 
 library built_value;
 
-export 'package:quiver/core.dart' show hashObjects;
-
 /// Implement this for a Built Value.
 ///
 /// Then use built_value_generator.dart code generation functionality to
@@ -59,3 +57,19 @@ abstract class Builder<V extends Built<V, B>, B extends Builder<V, B>> {
 //
 // Fields marked with this annotation are allowed to be null.
 const String nullable = 'nullable';
+
+/// For use by generated code in calculating hash codes. Do not use directly.
+int $jc(int hash, int value) {
+  // Jenkins hash "combine".
+  hash = 0x1fffffff & (hash + value);
+  hash = 0x1fffffff & (hash + ((0x0007ffff & hash) << 10));
+  return hash ^ (hash >> 6);
+}
+
+/// For use by generated code in calculating hash codes. Do not use directly.
+int $jf(int hash) {
+  // Jenkins hash "finish".
+  hash = 0x1fffffff & (hash + ((0x03ffffff & hash) << 3));
+  hash = hash ^ (hash >> 11);
+  return 0x1fffffff & (hash + ((0x00003fff & hash) << 15));
+}