+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add checking for correct type arguments for Built and Builder interfaces.
 - Generate empty constructor with semicolon instead of {}.
+- Use ArgumentError.notNull for null errors.
 
 ## 0.1.5